requirements.txt
flask==2.3.2
flask-cors==4.0.0
jinja2==3.1.2
transformers==4.40.1
sentence-transformers==2.3.1
torch==2.9.1
faiss-cpu
numpy
